Conflicto resuelto.

+++ TPE1/src/saveImages.m
@@ -1,4 +1,4 @@
 function ans = saveImages(noisy, corrected)
-	imwrite(noisy, gray(255), 'noisy.bmp');
-	imwrite(corrected, gray(255), 'corrected.bmp');
+	imwrite(uint8(noisy), gray(255), 'noisy.bmp');
+	imwrite(uint8(corrected), gray(255), 'corrected.bmp');
 endfunction
